refactor(select): extract row-mask types and IndexExprResult into lance-select
Moves the row-address mask types (`RowAddrMask`, `NullableRowAddrMask`,
`RowAddrTreeMap`, `RowIdMask`, `RowIdSet`, `NullableRowAddrSet`, the
`RowSetOps` trait, and the `bitmap_to_ranges` / `ranges_to_bitmap`
helpers) plus the certainty-tagged result types (`IndexExprResult` and
`NullableIndexExprResult`) — including their `Not` / `BitAnd` / `BitOr`
algebra and the `drop_nulls` collapser — into a new `lance-select`
crate. This sits below `lance-index` in the dependency graph (deps:
arrow-array, arrow-buffer, byteorder, deepsize, itertools, roaring,
lance-core) and gives consumers — benchmarks, prefilter / read-planner
code, and future per-range filter-pushdown work — a path to depend on
the mask substrate without pulling in `lance-index`.

Two new things, beyond the move:

* `impl Not for NullableIndexExprResult` — extracted from the inline
  match in `ScalarIndexExpr::evaluate_impl`. The body is identical
  (`Exact(!m)`, `AtMost↔AtLeast(!m)`); the evaluator now just calls
  `!result`.

* `index_expr_result_from_parts` / `serialize_index_expr_result` free
  functions on `lance-index::scalar::expression`, replacing the old
  inherent methods `IndexExprResult::{from_parts, serialize_to_arrow}`.
  The arrow wire format and its `INDEX_EXPR_RESULT_SCHEMA` constant
  stay in `lance-index` so `lance-select` doesn't have to take an
  arrow-schema dependency.

`From<SearchResult> for NullableIndexExprResult` likewise stays in
`lance-index` (where `SearchResult` is defined).

All 39 import sites for `lance_core::utils::mask::*` were updated to
`lance_select::*`. The `lance` / `lance-index` / `lance-table` crates
gained a `lance-select.workspace = true` dependency.

Verified:

  * lance-select unit tests: 97 passed (the mask + nullable suite)
  * lance-index --lib: 302 passed
  * lance-table --lib: 104 passed
  * lance --lib: 1672 passed
  * `cargo clippy --workspace --tests -- -D warnings`: clean
