+++ .bumpversion.toml
@@ -97,6 +97,11 @@ replace = 'lance-namespace-impls = {{ version = "={new_version}"'
 
 [[tool.bumpversion.files]]
 filename = "Cargo.toml"
+search = 'lance-select = {{ version = "={current_version}"'
+replace = 'lance-select = {{ version = "={new_version}"'
+
+[[tool.bumpversion.files]]
+filename = "Cargo.toml"
 search = 'lance-table = {{ version = "={current_version}"'
 replace = 'lance-table = {{ version = "={new_version}"'
