+++ rust/lance-core/src/utils.rs
@@ -11,7 +11,6 @@ pub mod cpu;
 pub mod deletion;
 pub mod futures;
 pub mod hash;
-pub mod mask;
 pub mod parse;
 pub mod path;
 pub mod tempfile;
