+++ rust/lance-index/src/frag_reuse.rs
@@ -8,8 +8,8 @@ use arrow_array::{Array, ArrayRef, PrimitiveArray, RecordBatch, UInt64Array};
 use async_trait::async_trait;
 use deepsize::{Context, DeepSizeOf};
 use itertools::Itertools;
-use lance_core::utils::mask::RowAddrTreeMap;
 use lance_core::{Error, Result};
+use lance_select::RowAddrTreeMap;
 use lance_table::format::pb::fragment_reuse_index_details::InlineContent;
 use lance_table::format::{ExternalFile, Fragment, pb};
 use roaring::{RoaringBitmap, RoaringTreemap};
