+++ rust/lance/src/index/scalar_logical.rs
@@ -9,11 +9,11 @@ use std::sync::Arc;
 use async_trait::async_trait;
 use deepsize::{Context, DeepSizeOf};
 use futures::future::try_join_all;
-use lance_core::utils::mask::NullableRowAddrSet;
 use lance_core::{Error, Result};
 use lance_index::metrics::MetricsCollector;
 use lance_index::scalar::{AnyQuery, CreatedIndex, ScalarIndex, SearchResult, UpdateCriteria};
 use lance_index::{Index, IndexType};
+use lance_select::NullableRowAddrSet;
 use lance_table::format::IndexMetadata;
 use roaring::RoaringBitmap;
 use serde_json::json;
